series last changes

+++ Models/seriesModel.php
@@ -8,22 +8,22 @@ class Series {
     private $title;
     private $platformid;
 
-    public function __construct($idSeries,$titleSeries,$idPlatform){
+    public function __construct($idSeries = null,$titleSeries = null,$idPlatform = null){
         $this->id = $idSeries;
         $this->title = $titleSeries;
         $this->platformid = $idPlatform; 
     }
 
     //Getter
-    public function getId(){
+    public function getId():int{
         return $this->id;
     }
 
-    public function getTitle(){
+    public function getTitle():string{
         return $this->title;
     }
 
-    public function getplatformid(){
+    public function getplatformid():int{
         return $this->platformid;
     }
 
@@ -43,8 +43,18 @@ class Series {
     // READ: Get all series
     public function getAll() {
         $mysqli = Database::getDbConnection();
+        // Get the error message from the database connection
+        if(is_string($mysqli)){
+            throw new Exception("Error al conectar con la base de datos: ". $mysqli);
+        }
+
         $query = "SELECT * FROM $this->table";
-        $result = $mysqli->query($query);
+        try {
+            $result = $mysqli->query($query);
+        }catch (Exception $error) {
+            throw new Exception("Error al tratar de hacer la consulta: " . $error->getMessage());
+        }
+        
         $listData = [];
 
         if ($result->num_rows > 0) {
@@ -60,25 +70,32 @@ class Series {
     public function create(){
         $serieCreated = false;
         $mysqli = Database::getDbConnection();
-
-        // Use prepared statement to prevent SQL injection
-        $stmt = $mysqli->prepare("SELECT title FROM $this->table WHERE title = ?");
-        $stmt->bind_param("s", $this->title);  // 's' denotes that we're passing a string
-
-        $stmt->execute();
-        $result = $stmt->get_result();
-
-        if ($result->num_rows <= 0) {
-
-            //if dont exits its going to create it
-            $insertStmt = $mysqli->prepare("INSERT INTO $this->table (title,platformid) VALUES (?,?)");
-            $insertStmt->bind_param("si", $this->title, $this->platformid);
-
-            if ($insertStmt->execute()) {
-                $serieCreated = true;  // Platform created successfully
-            }
+        // Get the error message from the database connection
+        if(is_string($mysqli)){
+            throw new Exception("Error al conectar con la base de datos: ". $mysqli);
         }
 
+        // Use prepared statement to prevent SQL injection
+        try{
+            $stmt = $mysqli->prepare("SELECT title FROM $this->table WHERE title = ?");
+            $stmt->bind_param("s", $this->title);  // 's' denotes that we're passing a string
+
+            $stmt->execute();
+            $result = $stmt->get_result();
+
+            if ($result->num_rows <= 0) {
+
+                //if dont exits its going to create it
+                $insertStmt = $mysqli->prepare("INSERT INTO $this->table (title,platformid) VALUES (?,?)");
+                $insertStmt->bind_param("si", $this->title, $this->platformid);
+
+                if ($insertStmt->execute()) {
+                    $serieCreated = true;  // Platform created successfully
+                }
+            }
+        }catch (Exception $error) {
+            throw new Exception("Error al tratar de hacer la consulta: " . $error->getMessage());
+        }
         $stmt->close();
         $mysqli->close();
 
@@ -87,13 +104,21 @@ class Series {
 
     public function getItem(){
         $mysqli = Database::getDbConnection();
+        // Get the error message from the database connection
+        if(is_string($mysqli)){
+            throw new Exception("Error al conectar con la base de datos: ". $mysqli);
+        }
+
         // Use prepared statement to prevent SQL injection
-        $stmt = $mysqli->prepare("SELECT * FROM $this->table WHERE id = ?");
-        $stmt->bind_param("i", $this->id);  // 's' denotes that we're passing a string
+        try {
+            $stmt = $mysqli->prepare("SELECT * FROM $this->table WHERE id = ?");
+            $stmt->bind_param("i", $this->id);  // 's' denotes that we're passing a string
 
-        $stmt->execute();
-        $result = $stmt->get_result();
-
+            $stmt->execute();
+            $result = $stmt->get_result();
+        }catch (Exception $error) {
+            throw new Exception("Error al tratar de hacer la consulta: " . $error->getMessage());
+        }
         //if exist a result
         if ($result->num_rows > 0){
             foreach ($result as $item){
@@ -107,20 +132,28 @@ class Series {
     public function update() {
         $seriesUpdated = false;
         $mysqli = Database::getDbConnection();
-    
-        // Prepare the update statement to prevent SQL injection
-        $stmt = $mysqli->prepare("UPDATE $this->table SET title=?,platformid = ?  WHERE id = ?");
-        $stmt->bind_param("sii", $this->title, $this->platformid,$this->id);  // 's' for string (name), 'i' for integer (id)
-        
-        // Execute the statement
-        if ($stmt->execute()) {
-            // Check if any row was affected (i.e., if the update was successful)
-            if ($stmt->affected_rows > 0) {
-                $seriesUpdated = true;
-            } else {
-                // No rows were updated (perhaps the id doesn't exist or name is unchanged)
-                $seriesUpdated = false;
+        try {
+            // Get the error message from the database connection
+            if(is_string($mysqli)){
+                throw new Exception("Error al conectar con la base de datos: ". $mysqli);
             }
+
+            // Prepare the update statement to prevent SQL injection
+            $stmt = $mysqli->prepare("UPDATE $this->table SET title=?,platformid = ?  WHERE id = ?");
+            $stmt->bind_param("sii", $this->title, $this->platformid,$this->id);  // 's' for string (name), 'i' for integer (id)
+            
+            // Execute the statement
+            if ($stmt->execute()) {
+                // Check if any row was affected (i.e., if the update was successful)
+                if ($stmt->affected_rows > 0) {
+                    $seriesUpdated = true;
+                } else {
+                    // No rows were updated (perhaps the id doesn't exist or name is unchanged)
+                    $seriesUpdated = false;
+                }
+            }
+        }catch (Exception $error) {
+            throw new Exception("Error al tratar de hacer la consulta: " . $error->getMessage());
         }
         $stmt->close();
         $mysqli->close();
@@ -131,22 +164,30 @@ class Series {
     public function delete(){
         $deleted = false;
         $mysqli = Database::getDbConnection();
+        // Get the error message from the database connection
+        if(is_string($mysqli)){
+            throw new Exception("Error al conectar con la base de datos: ". $mysqli);
+        }
 
         //prepate statement to Prevent sql injection
-        $stmt = $mysqli->prepare("DELETE FROM $this->table WHERE id = ?");
-        $stmt->bind_param("i", $this->id);
+        try {
+            $stmt = $mysqli->prepare("DELETE FROM $this->table WHERE id = ?");
+            $stmt->bind_param("i", $this->id);
 
-        //execute the statement
-        if ($stmt->execute()){
-            //check rows
-            if($stmt->affected_rows > 0){
-                $deleted = true; //record succesfully deleted
-            }else {
-                // No rows were updated (perhaps the id doesn't exist or name is unchanged)
-                $deleted = false;
+            //execute the statement
+            if ($stmt->execute()){
+                //check rows
+                if($stmt->affected_rows > 0){
+                    $deleted = true; //record succesfully deleted
+                }else {
+                    // No rows were updated (perhaps the id doesn't exist or name is unchanged)
+                    $deleted = false;
+                }
             }
+            // Close the statement and connection
+        }catch (Exception $error) {
+            throw new Exception("Error al tratar de hacer la consulta: " . $error->getMessage());
         }
-        // Close the statement and connection
         $stmt->close();
         $mysqli->close();
